v0.2.0: 5D tensors, 64-bit shapes, new views, Python-style indexing, scalar item(), MNIST example & accuracy

+++ include/znet/autograd/shape_ops.hpp
@@ -6,6 +6,6 @@ namespace znet {
 
 // Reduce `src` to `dst_shape` by summing across broadcast/extra dims.
 // Assumes row-major contiguous layout for now (storage_offset==0).
-Tensor sum_to_shape(const Tensor& src, const std::vector<int>& dst_shape);
+Tensor sum_to_shape(const Tensor& src, const std::vector<int64_t>& dst_shape);
 
 } // namespace znet
